rename newsroom.news to newsroom.wire
to be consistent with the ui
