+++ newsroom/templates/layout.html
@@ -29,7 +29,7 @@
 
     <nav class="main navbar justify-content-start">
         <div class="navbar-brand p-2">
-            <a class="nav-link text-white"  href="{{ url_for('news.index') }}">{{ config.SITE_NAME }} // {{ gettext('Wire') }}</a>
+            <a class="nav-link text-white"  href="{{ url_for('wire.index') }}">{{ config.SITE_NAME }} // {{ gettext('Wire') }}</a>
         </div>
 
         {% if session.get('name') %}
